pyproject.toml
[project]
name = "langgraph-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=26.3.1",
    "isort>=8.0.1",
    "langchain>=1.2.15",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.2.1",
    "langchain-tavily>=0.2.18",
    "langgraph>=1.1.9",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.2",
]
